Logging out and back in safety measures
REFACTOR: Made it so that all FlounderFests become invalidated upon server exit. Hopefully I'm able to get the persistent state thing working so that this isn't needed, but I'm not sure if I'll have time to do so.
FIX: Made it so that logging out and then logging into another world doesn't keep the FlounderFestHud enabled